Consulta de cantidad de productos

+++ android-project/FarmaceuticaTAES/app/src/main/java/farmaceutica/taes/farmaceutica/presentacion/controlador/Impl/ValoracionesProductoFragment.java
@@ -49,16 +49,14 @@ public class ValoracionesProductoFragment extends BaseFragment {
         textView_nombre.setText(prod.getNombre());
         textView_descripcion.setText(prod.getDescripcion());
         try {
-        textView_cantidad1.setText(Integer.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.NADA)));
-        textView_cantidad2.setText(Integer.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.POCO)));
-        textView_cantidad3.setText(Integer.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.MEDIO)));
-        textView_cantidad4.setText(Integer.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.MUCHO)));
+        textView_cantidad1.setText(Long.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.NADA)));
+        textView_cantidad2.setText(Long.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.POCO)));
+        textView_cantidad3.setText(Long.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.MEDIO)));
+        textView_cantidad4.setText(Long.toString(fachada.obtenerCantidadValoracionProducto(getActivity(),prod.getCodNacional(),ValoracionProducto.MUCHO)));
 
         } catch (Exception e) {
             e.printStackTrace();
         }
-
-
     }
 
     public static ValoracionesProductoFragment newInstance() {
